+++ src/main/java/com/project/covid/application/gateways/ApiNinjaClient.java
@@ -1,7 +1,7 @@
 package com.project.covid.application.gateways;
 
 import com.project.covid.domain.CountryDomain;
-import com.project.covid.infra.controllers.dtos.requests.CountryRequest;
+import com.project.covid.infra.controllers.mappers.requests.CountryRequest;
 
 public interface ApiNinjaClient {
