+++ src/main/java/com/project/covid/infra/client/impl/ApiNinjaClientImpl.java
@@ -1,17 +1,19 @@
 package com.project.covid.infra.client.impl;
 
 import com.project.covid.application.gateways.ApiNinjaClient;
-import com.project.covid.config.GsonConfig;
 import com.project.covid.domain.CountryDomain;
 import com.project.covid.infra.client.ApiNinjaClientFeign;
-import com.project.covid.infra.controllers.dtos.requests.CountryRequest;
+import com.project.covid.infra.config.GsonConfig;
+import com.project.covid.infra.controllers.mappers.requests.CountryRequest;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Component;
 
+@Component
 public class ApiNinjaClientImpl implements ApiNinjaClient {
-    private static final Logger logger = LoggerFactory.getLogger(ApiNinjaClientImpl.class);
 
+    private static final Logger logger = LoggerFactory.getLogger(ApiNinjaClientImpl.class);
 
     private final ApiNinjaClientFeign apiNinjaClientFeign;
     private final String token;
